Create RE“Management of Chain supply in Pharmacies” ADME.md
PHARMACEUTICAL SUPPLY CHAIN MANAGEMENT
Logistics involve the integration of inventory, warehousing, material handling, packing and transportation of prodcuts with security.
The fact that all the pharmaceuticals products and supplies are manufactured outside Chile, make imperative the optimal atenction to Goblal systems of supplier relationship and the expantion of supply chain over bounderies of India and China. Former supplier was in U.S; Europe; Colombia and Brazil.